chore: move sb decorator to separate file
