+++ .storybook/preview.js
@@ -1,7 +1,4 @@
-import { createContext, useContext, useState, useMemo, useCallback } from 'react'
-import { ThemeProvider, createTheme, CssBaseline, Switch, FormControlLabel, Box } from '@mui/material'
-import { Provider } from 'react-redux'
-import { noop } from 'lodash'
+import { themeSwitcher } from "./themeswitcher"
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
@@ -14,39 +11,6 @@ export const parameters = {
   layout: 'fullscreen'
 }
 
-const ThemeContext = createContext({
-  mode: 'light',
-  setDarkTheme: noop,
-  setLightTheme: noop,
-  toggleTheme: noop
-})
-
-export const ThemeContextProvider = ({ mode: $mode, children }) => {
-  const [mode, setMode] = useState($mode);
-  const setDarkTheme = () => setMode('dark');
-  const setLightTheme = () => setMode('light');
-  const toggleTheme = () => setMode(mode => mode === 'dark' ? 'dark' : 'light')
-  const theme = createTheme({ palette: { mode } })
-  const handleSwitchMode = useCallback(({ target: { checked } }) => {
-    if (checked) {
-      setDarkTheme()
-    } else setLightTheme()
-  }, [])
-  return <ThemeContext.Provider value={{ mode, setDarkTheme, setLightTheme, toggleTheme }}>
-    <ThemeProvider theme={theme}>
-      {children}
-      <Box position="fixed" sx={{ right: 30, bottom: 30 }}>
-        <FormControlLabel label="Dark Mode" control={<Switch checked={mode === 'dark'} onChange={handleSwitchMode} />}>
-        </FormControlLabel>
-      </Box>
-    </ThemeProvider>
-  </ThemeContext.Provider >
-}
-
 export const decorators = [
-  (Story) =>
-    <ThemeContextProvider mode="light">
-      <CssBaseline />
-      <Story />
-    </ThemeContextProvider>
+  themeSwitcher
 ]
